requirements.txt
flask==3.1.0
PyPDF2==3.0.1
nltk==3.9.1
requests==2.32.3
pillow==10.4.0
numpy==2.2.4
moviepy==2.1.2
python-dotenv==1.1.0
gunicorn==21.2.0
